% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{print.deltaMAN}
\alias{print.deltaMAN}
\title{Print object of class deltaMAN.}
\usage{
\method{print}{deltaMAN}(x, ...)
}
\arguments{
\item{x}{an object of class deltaMAN.}

\item{...}{optional arguments passed to print for other classes created in the deltaMAN package.
Currently, it supports \code{"digits"} (the significant digits to be used)
and \code{"tex"} (a logical value indicating whether to generate formatted LaTeX output or plain output).
Also, optional argument \code{"transpose"} is available for objects of class \code{"multiDelta"}.}
}
\value{
No return value. Results are printed on console.
}
\description{
\code{print} method for class \code{"deltaMAN"}.
}
