% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDeltaShinyApp.R
\name{runDeltaShinyApp}
\alias{runDeltaShinyApp}
\title{Run Delta Shiny App}
\usage{
runDeltaShinyApp()
}
\value{
No return value. A shiny app is launched.
}
\description{
Launch the Delta Shiny App to compute the Delta Measurement of Agreement for Nominal, from a user friendly interface.
The app also allows to download a report of results in .pdf or .tex (LaTeX) format. The latter format is compressed in a .zip file.
}
