% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{summary.delta}
\alias{summary.delta}
\title{Summary of Delta model}
\usage{
\method{summary}{delta}(object, fullReport = FALSE, digits = 4, tex = FALSE, ...)
}
\arguments{
\item{object}{and object of class \code{"delta"}}

\item{fullReport}{a logical value indicating whether to generate an exhaustive report of the results (TRUE) or not (FALSE, default).}

\item{digits}{an integer value indicating the significant digits to be used.}

\item{tex}{a logical value indicating whether to generate formatted LaTeX output (TRUE) or plain output (FALSE, default).}

\item{...}{further arguments passed to summary. No one else is currently available.}
}
\value{
No return value. Results are printed on console.
}
\description{
\code{summary} method for class \code{"delta"}.
This functions creates a report with the information contained in an object of class \code{"delta"}.
}
